requirements.txt
deepl>=1.0.1
python-dotenv>=0.17.0
recipe-scrapers==13.3.5
pdfkit==0.6.1
